5 and 1/2 6

+++ app/src/domain/posts/use_cases/delete_post.py
@@ -1,7 +1,14 @@
-from src.core.exceptions.database_exceptions import PostNotFoundById
-from src.core.exceptions.domain_exceptions import PostNotFoundByIdException
+import logging
+
+from src.core.exceptions.database_exceptions import PostNotFound
+from src.core.exceptions.domain_exceptions import (
+    PermissionDeniedException,
+    PostNotFoundByIdException,
+)
 from src.infrastructure.sqlite.database import database
 from src.infrastructure.sqlite.repositories.posts import PostRepository
+
+logger = logging.getLogger(__name__)
 
 
 class DeletePostUseCase:
@@ -9,11 +16,19 @@ class DeletePostUseCase:
         self._database = database
         self._repo = PostRepository()
 
-    async def execute(self, post_id: int) -> None:
-        try:
-            with self._database.session() as session:
-                self._repo.get_by_id(session, post_id)
-                self._repo.delete(session, post_id)
-                session.commit()
-        except PostNotFoundById as err:
-            raise PostNotFoundByIdException(id=post_id) from err
+    async def execute(self, post_id: int, current_user_id: int) -> None:
+        with self._database.session() as session:
+            try:
+                post = self._repo.get_by_id(session, post_id)
+            except PostNotFound as err:
+                error = PostNotFoundByIdException(post_id=post_id)
+                logger.error(error.detail)
+                raise error from err
+
+            if post.author_id != current_user_id:
+                error = PermissionDeniedException()
+                logger.error(error.detail)
+                raise error
+
+            self._repo.delete(session, post_id)
+            session.commit()
